# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Problem {
  type: string;
  title: string;
  status: number;
  detail: string;
  instance?: string;
}